# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Swagger",
        "description": "An API of cool stuffs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/empleados": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Obtener listado paginado de empleados",
                "operationId": "01e9d10f171ec0d88e4010f4876a0b65",
                "responses": {
                    "200": {
                        "description": "Lista paginada de empleados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "description": "Datos paginados",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Crear un nuevo empleado",
                "operationId": "8dbd970d554ec6064a0527ac54baa93f",
                "responses": {
                    "201": {
                        "description": "Empleado creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleados/{value}": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Buscar empleados filtrando por varios campos",
                "operationId": "23c3f40d896b2176c1d5c9ab823058e5",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Valor para filtrar empleados",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron empleados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "not found"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron departamentos con los criterios especificados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleados/{id}": {
            "put": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualizar un empleado existente",
                "operationId": "032112c4832b57bdfbf99a313486352d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empleado a actualizar",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Empleado actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Eliminar un empleado",
                "operationId": "cc4ed7cb13a8f08b869cb1b1775ec40f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empleado a eliminar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT prefijado con Bearer. Ejemplo: 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}